migrate translation extraction code to babel package
