+++ packages/docusaurus-babel/package.json
@@ -29,6 +29,8 @@
   "license": "MIT",
   "dependencies": {
     "@babel/core": "^7.23.3",
+    "@babel/generator": "^7.23.3",
+    "@babel/traverse": "^7.22.8",
     "@babel/plugin-syntax-dynamic-import": "^7.8.3",
     "@babel/plugin-transform-runtime": "^7.22.9",
     "@babel/preset-env": "^7.22.9",
@@ -36,10 +38,12 @@
     "@babel/preset-typescript": "^7.22.5",
     "@babel/runtime": "^7.22.6",
     "@babel/runtime-corejs3": "^7.22.6",
+    "@docusaurus/logger": "3.5.2",
     "@docusaurus/utils": "3.5.2",
     "babel-loader": "^9.1.3",
     "babel-plugin-dynamic-import-node": "^2.3.3",
-    "fs-extra": "^11.1.1"
+    "fs-extra": "^11.1.1",
+    "tslib": "^2.6.0"
   },
   "engines": {
     "node": ">=18.0"
