+++ packages/docusaurus-utils/src/globUtils.ts
@@ -10,8 +10,11 @@
 import path from 'path';
 import Micromatch from 'micromatch'; // Note: Micromatch is used by Globby
 import {addSuffix} from '@docusaurus/utils-common';
+import Globby from 'Globby';
+import {posixPath} from './pathUtils';
+
 /** A re-export of the globby instance. */
-export {default as Globby} from 'globby';
+export {Globby};
 
 /**
  * The default glob patterns we ignore when sourcing content.
@@ -84,4 +87,19 @@ export function createAbsoluteFilePathMatcher(
 
   return (absoluteFilePath: string) =>
     matcher(getRelativeFilePath(absoluteFilePath));
+}
+
+// Globby that fix Windows path patterns
+// See https://github.com/facebook/docusaurus/pull/4222#issuecomment-795517329
+export async function safeGlobby(
+  patterns: string[],
+  options?: Globby.GlobbyOptions,
+): Promise<string[]> {
+  // Required for Windows support, as paths using \ should not be used by globby
+  // (also using the windows hard drive prefix like c: is not a good idea)
+  const globPaths = patterns.map((dirPath) =>
+    posixPath(path.relative(process.cwd(), dirPath)),
+  );
+
+  return Globby(globPaths, options);
 }
