+++ packages/docusaurus/src/webpack/__tests__/base.test.ts
@@ -9,7 +9,6 @@ import {jest} from '@jest/globals';
 import path from 'path';
 import _ from 'lodash';
 import webpack from 'webpack';
-import * as utils from '@docusaurus/utils/lib/webpackUtils';
 import {posixPath} from '@docusaurus/utils';
 import {excludeJS, clientDir, createBaseConfig} from '../base';
 import {
@@ -135,20 +134,18 @@ describe('base webpack config', () => {
   });
 
   it('uses svg rule', async () => {
-    const isServer = true;
-    const fileLoaderUtils = utils.getFileLoaderUtils(isServer);
-    const mockSvg = jest.spyOn(fileLoaderUtils.rules, 'svg');
-    jest
-      .spyOn(utils, 'getFileLoaderUtils')
-      .mockImplementation(() => fileLoaderUtils);
-
-    await createBaseConfig({
+    const config = await createBaseConfig({
       props,
       isServer: false,
       minify: false,
       faster: DEFAULT_FASTER_CONFIG,
       configureWebpackUtils: await createTestConfigureWebpackUtils(),
     });
-    expect(mockSvg).toHaveBeenCalled();
+
+    const svgRule = (config.module?.rules ?? []).find((rule) => {
+      return rule && (rule as any).test.toString().includes('.svg');
+    });
+    expect(svgRule).toBeDefined();
+    expect(svgRule).toMatchSnapshot();
   });
 });
